Got the unit tests running again, apart from the two that were not working in the first place.  The changes to process_relationships and to sqlachemy itself were causing a double 'assign_mapper' call to cause issues.  Now, we basically defer calling assign_mapper until process_relationships in cases where there are defined relationships.
Also, I moved ActiveMapper to always use the default engine, as there were a
lot of hacks inside ActiveMapper to allow for engine swapping.  The use of the
default engine and the "global_connect" functionality significantly improves
the usability of ActiveMapper.

ActiveMapper will be getting a bit of a refactor/cleanup at some point in the
nearish future, as it has drifted a bit to complexity with the addition of some
features.  For now, this should do the trick!



git-svn-id: http://svn.sqlalchemy.org/sqlalchemy/trunk@1325 8cd8332f-0806-0410-a4b6-96f4b9520244