- compiler visit_label() checks a flag "within_order_by" and will render its own name and not its contained expression, if the dialect reports true for supports_simple_order_by_label. the flag is not propagated forwards, meant to closely mimic the syntax Postgres expects which is that only a simple name can be in the ORDER BY, not a more complex expression or function call with the label name embedded (mysql and sqlite support more complex expressions).
This further sets the standard for propigation of **kwargs within compiler, that we can't just send 
**kwargs along blindly to each XXX.process() call; whenever a **kwarg needs to propagate through, 
most methods will have to be aware of it and know when they should send it on forward and when not.
This was actually already the case with result_map as well.

The supports_simple_order_by dialect flag defaults to True but is conservatively explicitly set to
False on all dialects except SQLite/MySQL/Postgres to start.

[ticket:1068]


git-svn-id: http://svn.sqlalchemy.org/sqlalchemy/trunk@4999 8cd8332f-0806-0410-a4b6-96f4b9520244