- Added new ENUM type to the Postgresql dialect, which exists as a schema-level construct and extends the generic Enum type.  Automatically associates itself with tables and their parent metadata to issue the appropriate CREATE TYPE/DROP TYPE commands as needed, supports unicode labels, supports reflection.  [ticket:1511]
- MySQL ENUM now subclasses the new generic Enum type, and also handles
unicode values implicitly, if the given labelnames are unicode
objects.

- Added a new Enum generic type, currently supported on
Postgresql and MySQL.  Enum is a schema-aware object
to support databases which require specific DDL in 
order to use enum or equivalent; in the case of PG
it handles the details of `CREATE TYPE`, and on 
other databases without native enum support can 
support generation of CHECK constraints.
[ticket:1109] [ticket:1511]

- types documentation updates

- some cleanup on schema/expression docs


git-svn-id: http://svn.sqlalchemy.org/sqlalchemy/trunk@6435 8cd8332f-0806-0410-a4b6-96f4b9520244